Removed seemingly unnecessary hook.
There's a hook to decrement DR counter when enemies die in a map battle and another for animated battle. It seems the map battle hook also covers animated battles. Thus, we removed the hook for animated battles at 0x32808.

This should also remove the conflict this patch has with Leonarth's Less Annoying Fog.